# YJIT: Enhance the `String#<<` method substitution to handle integer codepoint values.