This PR extends YJIT's method substitution for `String#<<` to handle integer codepoints as well. If the string is `ASCII-8BIT` and the codepoint is a byte value, YJIT will dispatch to `rb_str_buf_cat_byte` as a fast path for working with binary strings. Otherwise, it'll dispatch to the general `rb_str_concat` just as `vm_opt_ltlt` would.

`rb_str_buf_cat_byte` currently works with both `ASCII-8BIT` and `US-ASCII`, but this YJIT side only optimizes for `ASCII-8BIT`. It could be extended easily enough with an additional comparison. The encoding indices for a handful of encodings, including both `ASCII-8BIT` and `US-ASCII` are fixed and sequential, so we could also do a range check. For the time being, I've omitted the handling of `US-ASCII`. I'd like to get feedback on the simplified PR and extend it with `US-ASCII` handling if needed (which I'm also not convinced we do).

Please advise if the mechanism I'm using to handle polymorphic dispatch is incorrect. We already have `jit_rb_str_concat` as a method substitution for `String#<<`, but it deliberately only handled string arguments. I could have merged the two, but that struck me as being complicated. However, I also don't know if it's fine to call between methods like this. And, it ends up duplicating some of the type checks to ensure we dispatch to the correct type depending on what we see at compile time. I was unsure on how to handle the runtime guards, so please pay extra attention to that.