@maximecb 2024-06-21
Hi Kevin.

Commenting to give you as much feedback as possible.

> This PR extends YJIT's method substitution for `String#<<` to handle integer codepoints as well. If the string is `ASCII-8BIT` and the codepoint is a byte value, YJIT will dispatch to `rb_str_buf_cat_byte` as a fast path for working with binary strings.

I think that the idea of this PR is a good one. It seems like a sensible specialization to implement.

I wrote some comments on the code itself.

Two things I would like to see in a PR like this:
- Benchmark results to know how it affects performance on protoboeuf and our headline benchmarks (if any impact)
- Maybe a dump of machine code generated when concatenating a few bytes in a loop just to see if we're doing anything obviously inefficient

> Please advise if the mechanism I'm using to handle polymorphic dispatch is incorrect. We already have `jit_rb_str_concat` as a method substitution for `String#<<`, but it deliberately only handled string arguments. I could have merged the two, but that struck me as being complicated. However, I also don't know if it's fine to call between methods like this. And, it ends up duplicating some of the type checks to ensure we dispatch to the correct type depending on what we see at compile time. I was unsure on how to handle the runtime guards, so please pay extra attention to that.

Seems sensible. Duplicating checks that occur at run-time would not be good but the checks you duplicated occur at compilation time. It makes sense to write the logic for this in a separate function to avoid ending up with a megafunction that is hard to follow.