@nirvdrum 2024-07-10
> An alternative to what you have is to make a new Rust/C function that essentially does the logic that you currently inline into every site (`if codepoint.is_fixnum() { rb_str_buf_cat_byte } else { rb_str_concat }`). That way you only generate one ccall at each site.

Okay. I think you were simplifying, but for completeness the check is really (in pseudo-code) `if codepoint.is_fixnum() && receiver.is_binary_string() && codepoint.is_byte()`. `rb_str_buf_cat_byte` does not have that logic. The caller is expected to call it only when those conditions hold (actually, `rb_str_buf_cat_byte` also works on US-ASCII strings, but I omitted that in YJIT) and there are assertions at the start of `rb_str_buf_cat_byte` that checks those conditions are held.

> If you write it in C, it could go into string.c and you can additionally avoid removing `static` from rb_str_buf_cat_byte().

I'll have to introduce a new function. I don't think updating `rb_str_buf_cat_byte` to do those checks is the right way to go. It would mean duplicating checks in the interpreter to derive information we already have.