@XrXr 2024-07-29
>  Initially Aaron had reported a big speedup on protoboeuf with your original patch. ...  I don't fully understand why the original patch had such a big speedup in the first place compared to this.


It looks like the speed-up was on the encoding benchmark and you ran the decoding benchmark.