@maximecb 2024-07-29
> > Initially Aaron had reported a big speedup on protoboeuf with your original patch. ...  I don't fully understand why the original patch had such a big speedup in the first place compared to this.
> 
> It looks like the speed-up was on the encoding benchmark and you ran the decoding benchmark. [#11032 (comment)](https://github.com/ruby/ruby/pull/11032#issuecomment-2218808415)

Mildly embarrassed 😅 Nevertheless, not ideal that it would slow down decode. 

It does seem slower on encode, too but we should probably run these benchmarks on an AWS machine as well:

```
master: ruby 3.4.0dev (2024-07-29T15:07:53Z master 86a762ce56) [arm64-darwin23]
cat: ruby 3.4.0dev (2024-07-26T19:34:15Z yjit-optimize-stri.. 3c3aba4119) [arm64-darwin23]

-----------------  -----------  ----------  --------  ----------  -----------  ----------
bench              master (ms)  stddev (%)  cat (ms)  stddev (%)  cat 1st itr  master/cat
protoboeuf         76.2         2.9         78.6      0.8         1.00         0.97      
protoboeuf-encode  89.8         1.2         92.6      1.4         0.96         0.97      
str_concat         43.8         4.8         44.0      4.0         0.96         0.99      
-----------------  -----------  ----------  --------  ----------  -----------  ----------
```

Also, @nirvdrum, have you rebased this PR on master recently? Just to make sure it's a fair comparison.